+++ src/extension.ts
@@ -1,11 +1,11 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
-import * as vscode from 'vscode';
-import { MyProvider } from './features/webview/providerClass';
+import { ExtensionContext, window } from 'vscode';
+import { DeltaBuddiesProvider } from './features/providers/DeltaBuddiesProvider';
 
 // This method is called when your extension is activated
 // Your extension is activated the very first time the command is executed
-export function activate(context: vscode.ExtensionContext) {
+export function activate(context: ExtensionContext) {
 
 	// Use the console to output diagnostic information (console.log) and errors (console.error)
 	// This line of code will only be executed once when your extension is activated
@@ -15,9 +15,9 @@ export function activate(context: vscode.ExtensionContext) {
 	// Now provide the implementation of the command with registerCommand
 	// The commandId parameter must match the command field in package.json
 
-    const provider = new MyProvider(context.extensionUri);
+    const provider = new DeltaBuddiesProvider(context.extensionUri);
 
-    const providerDisposable = vscode.window.registerWebviewViewProvider(
+    const providerDisposable = window.registerWebviewViewProvider(
         'mydeltabuddies.aquarium',
         provider
     );