+++ src/features/helpers/helpers.ts
@@ -1,0 +1,12 @@
+export function getNonce(): string {
+    const chars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+    let text = '';
+    for (let i = 0; i < 32; i++) {
+        text += chars.charAt(Math.floor(Math.random() * chars.length));
+    }
+    return text;
+}
+
+export function escapeJsonCharacters(charactersJson: string): string {
+    return charactersJson.replace(/"/g, '&quot;');
+}