+++ CHANGELOG.md
@@ -6,4 +6,10 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
-- Initial release
+## [0.1.0] - 2026-07-16
+
+- Add character animation using spritesheet
+- Move character across window according to animation
+- Implement random direction of walking, with random duration
+- Add idle state with random duration
+- Contain the character within the window boundaries